refactor(server): consolidate provider auth session manager
